requirements.txt
anyio==3.6.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-ninja==0.21.0
ecdsa==0.18.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.4
keycloak==3.1.3
pyasn1==0.4.8
pydantic==1.10.7
python-jose==3.3.0
python-keycloak-client==0.2.3
pytils==0.4.1
PyYAML==5.4.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
typing-extensions==4.5.0
urllib3==1.26.15
